Implement q_remove_head and q_remove_tail
Function q_remove_head first checks if the queue is NULL or empty.
If not, it get the first element using macro 'list_entry', removes it
from the queue using macro 'list_del_init', and returns the entry.

Function q_remove_tail also checks if the queue is NULL or empty first.
If not, it calls q_remove_head at head->prev->prev, this element is
always the previous node of tail if queue is not NULL or empty, then
reuse q_remove_head to remove next node of it.

Change-Id: I3529de374e5808c7672d8c6a2284aa175d3d5e54