Implement q_merge function to merge two list
This commit implements 'q_merge' function using a while loop to merge
each two queue in the chain until length of chain is one.
When merging pair of queues, the count of pair of queues is floor of
length / 2, move second queue to chain tail after merging each pair of
queues,
if current chain size is odd, the loop will remain last queue in the
chain and merge it in next round.

Change-Id: Ib640927eeee393043873c7009364b365f08d25ae