Implement queue ascend and descend function
These two function delete the nodes to keep list in ascend or descend
order.
The implement of 'q_ascend' is iterating the list from right, keep the
minimum value and compare it to each node to the left side, delete the
node if value is greater then minimum value.
The implement of 'q_descend' is the same to 'q_ascend' but keep maximum
value and delete the node that less then maximum value.

Change-Id: I2aeeb24db9a397b78467f88bdb7b24ecd9852811